change url to server and add json tag, move default function to default.go
